Extract non-handler functions to separate utils file
